// client/src/app/models/medico.ts
export interface Medico {
  id?: string;
  nombre?: string;
  apellido?: string;
  contrasenia?: string;
  direccion?: string;
  barrio?: string;
  idBarrio?: string;
  tipoID?: string;
  idTipoID?: string;
  universidad?: string;
  idUniversidad?: string;
  entidad?: string;
  idEntidad?: string;
}